+++ src/builtin.ml
@@ -823,6 +823,18 @@ counter C N :- trace.counter C N.|};
 
   MLData loc;
 
+  MLCode(Pred("loc.fields",
+    In(loc, "Loc",
+    Out(string, "File",
+    Out(int, "StartChar",
+    Out(int, "StopChar",
+    Out(int, "Line",
+    Out(int, "LineStartsAtChar",
+    Easy "Decomposes a loc into its fields")))))),
+  (fun { source_name; source_start; source_stop; line; line_starts_at; } _ _ _ _ _ ~depth:_ ->
+     !: source_name +! source_start +! source_stop +! line +! line_starts_at )),
+  DocAbove);
+
   LPDoc "== Regular Expressions =====================================";
 
   MLCode(Pred("rex.match",