+++ CHANGES.md
@@ -4,6 +4,7 @@
   - Fix `std.findall` is now calling a builtin which is able to produce
     solutions which contain eigenvariables and uvars as well.
   - `loc` is now printed using `/` as a path separator also on Windows
+  - `loc.fields` to project a `loc` into the file, line, etc...
 
 # v1.13.5 (May 2021)